+++ R/colcomb2class.R
@@ -3,7 +3,7 @@
 #' Combine columns from qdap classes or a \code{data.frame}.
 #' 
 #' @param dataframe A dataframe or qdap class (e.g., 
-#' "termco", "question_type", "pos_by", "character_table").
+#' \code{termco}, \code{question_type}, \code{pos_by}, \code{character_table}).
 #' @param combined.columns A list of named vectors of the colnames/indexes 
 #' of the numeric columns to be combined (summed).  If a vector is unnamed a 
 #' name will be assigned.